requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
langchain==0.1.11
langchain-community==0.0.27
langchain-core==0.1.30
langchain-openai==0.0.5
pydantic==2.5.3
python-multipart==0.0.9
SQLAlchemy==2.0.25
python-dotenv==1.0.0
pypdf==4.0.1
tiktoken==0.5.2
transformers==4.38.1
faiss-cpu==1.7.4
locust==2.20.1
anthropic==0.21.1
openai==1.11.1
pandas==2.1.3
matplotlib==3.8.2
pymongo==4.6.1
markdown==3.5.1
pytest==7.4.3
pytest-cov==4.1.0
httpx==0.25.2
backoff==2.2.1
chromadb==0.4.22
pymupdf==1.23.19
PyPDF2==3.0.1
mistune==3.0.1 